Add t0 configs: metadata, port (without serdes)
Add two T0 configurations, metadata and port (without serdes) base on the config_db.json configuration file.
Define the new setup method in T0 broadcom.
Add a config file for the s6000_s1220_r0
Refactor the method for abstract the info from config.ini, make the key from the name to the index, which can map the the port index directly.

Todo:
Add the serdes config, which needs to read the config file and parse the info.

Test done:
Tested on the dut.